+++ Sources/SmileID/Classes/DocumentVerification/View/DocumentCaptureScreen.swift
@@ -94,6 +94,19 @@ public struct DocumentCaptureScreen: View {
                 cameraManager: viewModel.cameraManager,
                 onCaptureClick: viewModel.captureDocument
             )
+            .alert(item: $viewModel.unauthorizedAlert) { alert in
+                Alert(
+                    title: Text(alert.title),
+                    message: Text(alert.message ?? ""),
+                    primaryButton: .default(
+                        Text(SmileIDResourcesHelper.localizedString(for: "Camera.Unauthorized.PrimaryAction")),
+                        action: {
+                            viewModel.openSettings()
+                        }
+                    ),
+                    secondaryButton: .cancel()
+                )
+            }
         }
     }
 }