+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Release Notes
+## 10.2.2
+
+#### Fixed
+*  Remove force unwrapping for getting files from storage
+
+## 10.2.1
+
+#### Added
+*  Handle unauthorized camera status
 
 ## 10.2.0